feat(exercise): Implement soft delete for related exercises
- Added `IsDeleted` flag check to avoid issues when exercises have related entries.
- Introduced `IsExerciseRelated` method for determining if an exercise has associated entries before deletion.

+++ MixxFit.API/Features/Exercises/DeleteExercise/DeleteExerciseHandler.cs
@@ -17,13 +17,25 @@ public class DeleteExerciseHandler(AppDbContext context) : IHandler
             .FirstOrDefaultAsync(e => e.Id == id && e.UserId == userId, cancellationToken);
 
         if (exercise is null)
-        {
             return Result.Failure(GeneralError.NotFound("Exercise"));
+        
+        if (await IsExerciseRelated(id))
+        {
+            exercise.IsDeleted = true;
+            await context.SaveChangesAsync(cancellationToken);
+            return Result.Success();
         }
 
         context.Exercises.Remove(exercise);
         await context.SaveChangesAsync(cancellationToken);
 
         return Result.Success();
+    }
+
+    private async Task<bool> IsExerciseRelated(int exerciseId)
+    {
+        return await context.Exercises
+            .Where(e => e.Id == exerciseId)
+            .AnyAsync(e => e.ExerciseEntries.Count > 0);
     }
 }
